refactor(device_map): split package to satisfy 200-LOC module cap.
Move GPU discovery, layout, parsing, resolution, and status helpers into
focused submodules under acestep/device_map/ with a stable public facade.
